+++ packages/client/hmi-client/src/components/workflow/ops/optimize-ciemss/tera-static-intervention-policy-group.vue
@@ -141,7 +141,7 @@ import {
 } from '@/components/workflow/ops/optimize-ciemss/optimize-ciemss-operation';
 import TeraSignalBars from '@/components/widgets/tera-signal-bars.vue';
 import teraTimestepCalendar from '@/components/widgets/tera-timestep-calendar.vue';
-import { getCalendarSettingsFromModel } from '@/utils/date';
+import { getCalendarSettingsFromModel } from '@/services/model';
 
 const props = defineProps<{
 	model: Model;
